ic_uber/interface/ic_uber/interface/requirements.txt
Django == 3.1
django-environ == 0.4.5
gunicorn == 20.0.4
psycopg2-binary==2.8.6
whitenoise == 5.2.0
django-heroku==0.0.0



dj-database-url==0.5.0
django-extensions==3.1.0
djangorestframework==3.12.2
python-decouple==3.4


django-bootstrap-datepicker-plus==3.0.5
